--- pkg/objectcache/networkneighborhoodcache_interface.go
@@ -1,28 +1,0 @@
-package objectcache
-
-import (
-	containercollection "github.com/inspektor-gadget/inspektor-gadget/pkg/container-collection"
-	"github.com/kubescape/storage/pkg/apis/softwarecomposition/v1beta1"
-)
-
-type NetworkNeighborhoodCache interface {
-	GetNetworkNeighborhood(containerID string) *v1beta1.NetworkNeighborhood
-	GetNetworkNeighborhoodState(containerID string) *ProfileState
-	ContainerCallback(notif containercollection.PubSubEvent)
-}
-
-var _ NetworkNeighborhoodCache = (*NetworkNeighborhoodCacheMock)(nil)
-
-type NetworkNeighborhoodCacheMock struct {
-}
-
-func (nn *NetworkNeighborhoodCacheMock) GetNetworkNeighborhood(_ string) *v1beta1.NetworkNeighborhood {
-	return nil
-}
-
-func (nn *NetworkNeighborhoodCacheMock) ContainerCallback(_ containercollection.PubSubEvent) {
-}
-
-func (nn *NetworkNeighborhoodCacheMock) GetNetworkNeighborhoodState(_ string) *ProfileState {
-	return nil
-}
